Associate user model with post model in one-to-many relationship

+++ Feeds/Controllers/PostController.cs
@@ -1,7 +1,10 @@
-﻿using Feeds.Data;
+﻿using System.Diagnostics;
+using System.Security.Claims;
+using Feeds.Data;
 using Feeds.Models;
 using Feeds.Repositories;
 using Feeds.Utilities;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Feeds.Controllers;
@@ -11,22 +14,26 @@ public class PostController : Controller
     private readonly ApplicationDbContext _dbContext;
     private readonly IWebHostEnvironment _webHostEnvironment;
     private readonly IUnitOfWork _unitOfWork;
+    private readonly IHttpContextAccessor _httpContextAccessor;
 
     public PostController(ApplicationDbContext dbContext, IUnitOfWork unitOfWork,
-        IWebHostEnvironment webHostEnvironment)
+        IWebHostEnvironment webHostEnvironment, IHttpContextAccessor httpContextAccessor)
     {
         _dbContext = dbContext;
         _webHostEnvironment = webHostEnvironment;
         _unitOfWork = unitOfWork;
+        _httpContextAccessor = httpContextAccessor;
     }
 
+    [AllowAnonymous]
     [Route("posts/")]
     public IActionResult Index()
     {
-        var postList = _unitOfWork.PostRepository.GetAll();
+        var postList = _unitOfWork.PostRepository.GetAll("ApplicationUser");
         return View(postList.OrderBy(p => p.Title));
     }
 
+    [AllowAnonymous]
     [Route("posts/{year:int}/{month:int}/{day:int}/{slug}")]
     public IActionResult Get(int? year, int? month, int? day, string? slug)
     {
@@ -36,10 +43,11 @@ public class PostController : Controller
         }
 
         Post post = _unitOfWork.PostRepository.Get(p =>
-            p.Slug == slug && p.CreatedOn.Year == year && p.CreatedOn.Month == month && p.CreatedOn.Day == day);
+            p.Slug == slug && p.CreatedOn.Year == year && p.CreatedOn.Month == month && p.CreatedOn.Day == day, "ApplicationUser");
         return View(post);
     }
 
+    [Authorize]
     [HttpGet]
     [Route("posts/create")]
     public IActionResult Create()
@@ -47,6 +55,7 @@ public class PostController : Controller
         return View();
     }
 
+    [Authorize]
     [HttpPost]
     [Route("posts/create")]
     public IActionResult Create(Post post)
@@ -54,33 +63,43 @@ public class PostController : Controller
         IFormFile file = Request.Form.Files.FirstOrDefault();
         var targetPath = @"images/posts";
         FileManagementUtility fileManagementUtility = new FileManagementUtility(_webHostEnvironment);
-       
+        
+        var userId = _httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier);
         if (ModelState.IsValid)
         {
             if (_dbContext.Posts.Any(p =>
-                    p.Slug == post.Slug))
+                    p.Slug == post.Slug && p.CreatedOn.Year == DateTime.Now.Year &&
+                    p.CreatedOn.Month == DateTime.Now.Month && p.CreatedOn.Day == DateTime.Now.Day))
             {
                 ModelState.AddModelError("", "Post title can not be duplicate");
                 return View();
             }
+            post.ApplicationUserId = userId;
             post.Image = fileManagementUtility.UploadFile(file, targetPath);
             _unitOfWork.PostRepository.Add(post);
             _unitOfWork.Save();
             return RedirectToAction(nameof(Index));
         }
-
         return View();
     }
 
+    [Authorize]
     [Route("posts/{year:int}/{month:int}/{day:int}/{slug}/edit")]
     public IActionResult Edit(int? year, int? month, int? day, string? slug)
     {
+        var userId = _httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier);
+        
         if (slug == null) return NotFound();
         Post post = _unitOfWork.PostRepository.Get(p =>
             p.Slug == slug && p.CreatedOn.Year == year && p.CreatedOn.Month == month && p.CreatedOn.Day == day);
-        return View(post);
+        if (post.ApplicationUserId == userId)
+        {
+            return View(post);
+        }
+        return RedirectToAction(nameof(Index));
     }
 
+    [Authorize]
     [HttpPost]
     [Route("posts/{year:int}/{month:int}/{day:int}/{slug}/edit")]
     public IActionResult Edit(Post post)
@@ -96,6 +115,7 @@ public class PostController : Controller
                 post.Image = fileManagementUtility.UploadFile(file, targetPath);
             }
 
+            post.ApplicationUserId = post.ApplicationUserId;
             post.Image = post.Image;
             _unitOfWork.PostRepository.Update(post);
             _unitOfWork.Save();
@@ -105,9 +125,11 @@ public class PostController : Controller
                     slug = post.Slug, year = post.CreatedOn.Year, month = post.CreatedOn.Month, day = post.CreatedOn.Day
                 });
         }
+
         return View();
     }
 
+    [Authorize]
     [HttpPost]
     public IActionResult Delete(int? id)
     {
@@ -117,8 +139,12 @@ public class PostController : Controller
         }
 
         Post post = _unitOfWork.PostRepository.Get(p => p.Id == id);
-        _unitOfWork.PostRepository.Remove(post);
-        _unitOfWork.Save();
+        var userId = _httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier);
+        if (post.ApplicationUserId == userId)
+        {
+            _unitOfWork.PostRepository.Remove(post);
+            _unitOfWork.Save();
+        }
         return RedirectToAction(nameof(Index));
     }
 }
